+++ prisma/migrations/20221011175118_update_http_response_model/migration.sql
@@ -1,0 +1,2 @@
+-- DropIndex
+DROP INDEX "HTTPResponse_environmentId_key";