+++ src/main/utils/logRotation.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
 import path from 'path';
 import * as fs from 'fs';
 import log from 'electron-log';
@@ -9,25 +10,30 @@ import getAppDataFolder from './fsUtils';
  * @since 0.1.0
  */
 export default function rotateLogFile(): void {
-  const todayDateFormat = new Date()
-    .toLocaleDateString('pt')
-    .split('/')
-    .reverse()
-    .join('-');
-  const filePath = path.resolve(
-    getAppDataFolder(),
-    'logs',
-    isDevelopment ? 'app.dev.log' : 'app.log'
-  );
-
-  if (!fs.existsSync(filePath.replace('.log', `_${todayDateFormat}.log`))) {
-    log.info('This log file needs rotation and will be archived');
-
-    fs.renameSync(
-      filePath,
-      filePath.replace('.log', `_${todayDateFormat}.log`)
+  try {
+    const todayDateFormat = new Date()
+      .toLocaleDateString('pt')
+      .split('/')
+      .reverse()
+      .join('-');
+    const filePath = path.resolve(
+      getAppDataFolder(),
+      'logs',
+      isDevelopment ? 'app.dev.log' : 'app.log'
     );
 
-    log.info('Previous log file has been archived due to file rotation');
+    if (!fs.existsSync(filePath.replace('.log', `_${todayDateFormat}.log`))) {
+      log.info('This log file needs rotation and will be archived');
+
+      fs.renameSync(
+        filePath,
+        filePath.replace('.log', `_${todayDateFormat}.log`)
+      );
+
+      log.info('Previous log file has been archived due to file rotation');
+    }
+  } catch (e: any) {
+    log.error('Could not rotate log file: ');
+    log.error(e.stack);
   }
 }